Section 9

+++ backend/routes/posts.js
@@ -36,7 +36,8 @@ router.post(
     const post = new Post({
       title: req.body.title,
       content: req.body.content,
-      imagePath: url + '/images/' + req.file.filename
+      imagePath: url + '/images/' + req.file.filename,
+      creator: req.userData.userId
     });
     post.save().then((createdPost) => {
       res.status(201).json({
@@ -64,13 +65,16 @@ router.put(
       _id: req.body.id,
       title: req.body.title,
       content: req.body.content,
-      imagePath: imagePath
+      imagePath: imagePath,
+      creator: req.userData.userId
     });
-    console.log(post);
-    Post.updateOne({ _id: req.params.id }, post)
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post)
     .then(result => {
-      console.log(result);
-      res.status(200).json({ message: 'Update successful!' });
+      if(result.nModified > 0) {
+        res.status(200).json({ message: 'Update successful!' });
+      } else {
+        res.status(401).json({ message: 'Not Authorized!' });
+      }
     });
 });
 
@@ -115,10 +119,14 @@ router.delete(
   '/:id',
   checkAuth,
   (req, res, next) => {
-    Post.deleteOne({ _id: req.params.id })
+    Post.deleteOne({ _id: req.params.id, creator: req.userData.userId })
     .then((result) => {
       console.log(result);
-      res.status(200).json({ message: 'Post deleted!' });
+      if(result.deletedCount > 0) {
+        res.status(200).json({ message: 'Post deleted!' });
+      } else {
+        res.status(401).json({ message: 'Not authorized!' });
+      }
     });
 });
